requirements.txt
# torch
# torchvision
timm==0.4.12
Pillow
blobfile
mypy
numpy
matplotlib
pytest
requests
einops
tensorboardX
# deepspeed==0.4.0
scipy
tqdm
six
opencv-python
lpips
# cd pytorch-fid; python setup.py install
# prettytable